requirements.txt
requests
pymongo
pandas
numpy
seaborn
certifi
python-dotenv
pymongo[srv]==3.6 
scikit-learn
matplotlib
statsmodels 
tensorflow 
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart
##-e .